refactor(build): split pre_pr frontmatter helpers by cohesion
Extract the frontmatter parser and DESIGN-REVIEW validator from
pre_pr.py (1401 lines) into sibling modules:

- yaml_utils.py: _parse_yaml_frontmatter (minimal, no-dependency parser).
- validate_design_review.py: validate_design_review_frontmatter plus its
  status/priority/blocking constants.

pre_pr.py re-exports both so they stay importable from scripts.validation.pre_pr
for tests and callers. Drops pre_pr.py to 1281 lines; further decomposition of
the all-PR-blocking runner is out of scope here (higher blast radius).

Also repoints the canonical-source citations in spec_contradiction.py from
pre_pr.py:_parse_yaml_frontmatter to yaml_utils.py:_parse_yaml_frontmatter,
which is where the parser now lives (.claude/rules/canonical-source-mirror.md).

Behavior-preserving. Refs #2223
